shape: extract recognition core to `aver::analysis::shape` (#232 stage 1)
First mechanical step of issue #232 (0.23 "Shape"). The per-fn shape
recognition primitives — `Facts` struct, AST walker, `extract_facts`,
multi-label `classify`, `primary_label`, call-graph SCC computation,
`ARCHETYPE_ORDER`, `type_is_result` — move out of
`aver::diagnostics::shape` into a new `aver::analysis::shape` tier.

Zero behavior change:

- `diagnostics::shape` keeps the recognition symbols re-exported with
  `pub use`, so existing consumers (CLI / LSP / research test) keep
  compiling without code changes.
- ModuleShape vector, Kind, Layer fingerprints, verify report,
  histogram, CorpusEntry walker, render_text / render_json / corpus
  renderers, aver.toml `[[shape.layer]]` / `[[shape.expected]]`
  bridges — all stay in `diagnostics::shape`. That's the presentation
  layer the CLI sees; stage 2 trims it to a facade.

Tests: `shape_spec` (14/14), `shape_lint_cli_spec` (3/3), and the
`research_archetype_clustering` ignored corpus survey all pass
unchanged.

Why the split (recap from issue #232):

- recognition (this module, `analysis::shape`) — typed facts over HIR
- meaning (`codegen::proof_lower`) — which patterns justify which
  ProofStrategy
- presentation (`diagnostics::shape`) — renderer / config / layer
  fingerprints / corpus walker / `aver shape` CLI

Stage 2 (next PR) turns the diagnostics module into a thin renderer
that imports from `analysis::shape` directly, instead of carrying the
core itself.
