+++ src/diagnostics/shape.rs
@@ -1,411 +1,31 @@
-//! `aver shape` — static module-shape analyzer.
+//! `aver shape` — module-shape presentation layer.
 //!
-//! Walks the resolved AST and emits:
+//! Stage 1 of issue #232 (0.23 "Shape") split this file into two
+//! tiers:
 //!
-//! - per-fn archetype classification (extracted from #104 research; see
-//!   `tests/research_archetype_clustering.rs` for the corpus-survey
-//!   harness that still uses these primitives)
-//! - per-module `ModuleShape` 5-dim vector (purity / entry / state_shape /
-//!   type_surface / api_shape) + derived `Kind` label
-//! - verify-block coverage report
-//! - archetype histogram + nearest-Layer fingerprint with confidence
+//! - **Recognition** (`aver::analysis::shape`) — per-fn facts walk,
+//!   archetype classifier, call-graph SCC. Stable typed primitives.
+//! - **Presentation** (this module) — `ModuleShape` vector + `Kind` +
+//!   verify report + histogram + Layer fingerprint distance + corpus
+//!   walker + renderer + `aver.toml [[shape.layer]]` / `[[shape.expected]]`
+//!   bridges. Everything the CLI / LSP / JSON consumers see.
 //!
-//! Layer fingerprints default to the built-in v0 table (see
-//! `builtin_v0_layer_fingerprints`). Projects override via
-//! `[[shape.layer]]` in `aver.toml`; `[[shape.expected]]` arms `--lint`.
+//! The recognition primitives are re-exported below so existing
+//! callers (and the research test) keep working through this module
+//! during the transition; downstream stages will migrate them to
+//! import directly from `aver::analysis::shape`.
 
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::path::Path;
 
 use crate::ast::TopLevel;
-use crate::ir::hir::{
-    ResolvedCallee, ResolvedCtor, ResolvedExpr, ResolvedFnDef, ResolvedMatchArm, ResolvedPattern,
-    ResolvedStmt, ResolvedStrPart, ResolvedTopLevel,
-};
+use crate::ir::hir::{ResolvedFnDef, ResolvedTopLevel};
 use crate::ir::{FnId, PipelineConfig, TypecheckMode};
 use crate::types::Type;
 
-// ─── Per-fn AST facts ────────────────────────────────────────────────────────
-
-#[derive(Debug, Default)]
-pub struct Facts {
-    pub calls_to: HashSet<FnId>,
-    pub tail_calls: HashSet<FnId>,
-    pub builtin_calls: Vec<String>,
-    pub ctor_constructs: Vec<ResolvedCtor>,
-    pub ctor_match_patterns: usize,
-    pub other_match_patterns: usize,
-    pub has_match: bool,
-    pub has_error_prop: bool,
-    pub record_creates: usize,
-    pub last_stmt_is_match: bool,
-    pub only_stmt_is_literal: bool,
-    pub body_stmt_count: usize,
-    pub has_interp_str: bool,
-    pub string_builtin_calls: usize,
-    pub has_match_with_err_arm: bool,
-}
-
-fn walk_expr(e: &ResolvedExpr, facts: &mut Facts) {
-    match e {
-        ResolvedExpr::Literal(_) | ResolvedExpr::Ident(_) | ResolvedExpr::Resolved { .. } => {}
-        ResolvedExpr::Attr(inner, _) => walk_expr(&inner.node, facts),
-        ResolvedExpr::Call(callee, args) => {
-            match callee {
-                ResolvedCallee::Fn(id) => {
-                    facts.calls_to.insert(*id);
-                }
-                ResolvedCallee::Builtin(name) => {
-                    if name.starts_with("String.") {
-                        facts.string_builtin_calls += 1;
-                    }
-                    facts.builtin_calls.push(name.clone());
-                }
-                ResolvedCallee::Intrinsic(_) => {}
-                ResolvedCallee::LocalSlot { .. } => {}
-                ResolvedCallee::Unresolved { callee } => walk_expr(&callee.node, facts),
-            }
-            for a in args {
-                walk_expr(&a.node, facts);
-            }
-        }
-        ResolvedExpr::BinOp(_, a, b) => {
-            walk_expr(&a.node, facts);
-            walk_expr(&b.node, facts);
-        }
-        ResolvedExpr::Neg(inner) => walk_expr(&inner.node, facts),
-        ResolvedExpr::Match { subject, arms } => {
-            facts.has_match = true;
-            walk_expr(&subject.node, facts);
-            for arm in arms {
-                count_arm_pattern(&arm.pattern, facts);
-                walk_match_arm(arm, facts);
-            }
-        }
-        ResolvedExpr::Ctor(c, args) => {
-            facts.ctor_constructs.push(c.clone());
-            for a in args {
-                walk_expr(&a.node, facts);
-            }
-        }
-        ResolvedExpr::ErrorProp(inner) => {
-            facts.has_error_prop = true;
-            walk_expr(&inner.node, facts);
-        }
-        ResolvedExpr::InterpolatedStr(parts) => {
-            facts.has_interp_str = true;
-            for p in parts {
-                if let ResolvedStrPart::Parsed(e) = p {
-                    walk_expr(&e.node, facts);
-                }
-            }
-        }
-        ResolvedExpr::List(xs) | ResolvedExpr::Tuple(xs) => {
-            for x in xs {
-                walk_expr(&x.node, facts);
-            }
-        }
-        ResolvedExpr::MapLiteral(pairs) => {
-            for (k, v) in pairs {
-                walk_expr(&k.node, facts);
-                walk_expr(&v.node, facts);
-            }
-        }
-        ResolvedExpr::RecordCreate { fields, .. } => {
-            facts.record_creates += 1;
-            for (_, v) in fields {
-                walk_expr(&v.node, facts);
-            }
-        }
-        ResolvedExpr::RecordUpdate { base, updates, .. } => {
-            walk_expr(&base.node, facts);
-            for (_, v) in updates {
-                walk_expr(&v.node, facts);
-            }
-        }
-        ResolvedExpr::TailCall { target, args } => {
-            facts.tail_calls.insert(*target);
-            facts.calls_to.insert(*target);
-            for a in args {
-                walk_expr(&a.node, facts);
-            }
-        }
-        ResolvedExpr::IndependentProduct(xs, _) => {
-            for x in xs {
-                walk_expr(&x.node, facts);
-            }
-        }
-    }
-}
-
-fn count_arm_pattern(p: &ResolvedPattern, facts: &mut Facts) {
-    match p {
-        ResolvedPattern::Ctor(ctor, _) => {
-            facts.ctor_match_patterns += 1;
-            if matches!(
-                ctor,
-                ResolvedCtor::Builtin(crate::ir::hir::BuiltinCtor::ResultErr)
-            ) {
-                facts.has_match_with_err_arm = true;
-            }
-        }
-        _ => facts.other_match_patterns += 1,
-    }
-}
-
-fn walk_match_arm(arm: &ResolvedMatchArm, facts: &mut Facts) {
-    walk_expr(&arm.body.node, facts);
-}
-
-pub fn extract_facts(fd: &ResolvedFnDef) -> Facts {
-    let mut facts = Facts::default();
-    let stmts = fd.body.stmts();
-    facts.body_stmt_count = stmts.len();
-    for stmt in stmts {
-        match stmt {
-            ResolvedStmt::Binding { value, .. } => walk_expr(&value.node, &mut facts),
-            ResolvedStmt::Expr(value) => walk_expr(&value.node, &mut facts),
-        }
-    }
-    if let Some(last) = stmts.last() {
-        let expr = match last {
-            ResolvedStmt::Binding { value, .. } => &value.node,
-            ResolvedStmt::Expr(value) => &value.node,
-        };
-        facts.last_stmt_is_match = matches!(expr, ResolvedExpr::Match { .. });
-    }
-    facts.only_stmt_is_literal = stmts.len() == 1 && {
-        let expr = match &stmts[0] {
-            ResolvedStmt::Binding { value, .. } => &value.node,
-            ResolvedStmt::Expr(value) => &value.node,
-        };
-        matches!(
-            expr,
-            ResolvedExpr::Literal(_)
-                | ResolvedExpr::List(_)
-                | ResolvedExpr::Tuple(_)
-                | ResolvedExpr::MapLiteral(_)
-                | ResolvedExpr::RecordCreate { .. }
-                | ResolvedExpr::Ctor(_, _)
-        )
-    };
-    facts
-}
-
-// ─── Per-fn classification ───────────────────────────────────────────────────
-
-pub fn classify(fd: &ResolvedFnDef, facts: &Facts, scc: &HashSet<FnId>) -> Vec<&'static str> {
-    let mut labels = Vec::new();
-
-    let self_call = facts.calls_to.contains(&fd.fn_id) || facts.tail_calls.contains(&fd.fn_id);
-    if self_call {
-        labels.push("structural-recursion");
-    }
-    if scc.contains(&fd.fn_id) {
-        labels.push("scc-mutual");
-    }
-
-    if facts.last_stmt_is_match {
-        if facts.ctor_match_patterns >= facts.other_match_patterns && facts.ctor_match_patterns > 0
-        {
-            labels.push("match-dispatcher");
-        } else {
-            labels.push("match-on-value");
-        }
-    }
-
-    if !fd.effects.is_empty() {
-        let total_calls = facts.calls_to.len() + facts.builtin_calls.len();
-        if total_calls >= 2 {
-            labels.push("orchestration");
-        } else {
-            labels.push("effectful-leaf");
-        }
-    }
-
-    let is_result_ret = type_is_result(&fd.return_type);
-    if is_result_ret && facts.has_error_prop {
-        labels.push("pipeline-result");
-    } else if is_result_ret && facts.has_match_with_err_arm {
-        labels.push("manual-result-adapter");
-    }
-
-    let is_string_ret = matches!(&fd.return_type, Type::Named { name, .. } if name == "String");
-    if is_string_ret
-        && fd.effects.is_empty()
-        && (facts.has_interp_str || facts.string_builtin_calls >= 1)
-        && (facts.body_stmt_count >= 2 || facts.has_interp_str)
-    {
-        labels.push("renderer-formatter");
-    }
-
-    if facts.body_stmt_count == 1
-        && (!facts.ctor_constructs.is_empty() || facts.record_creates > 0)
-        && !facts.has_match
-    {
-        labels.push("constructor-wrapper");
-    }
-
-    if fd.params.is_empty()
-        && facts.calls_to.is_empty()
-        && facts.builtin_calls.is_empty()
-        && fd.effects.is_empty()
-        && facts.only_stmt_is_literal
-    {
-        labels.push("data-as-function");
-    }
-
-    if facts.body_stmt_count == 1
-        && !facts.has_match
-        && !self_call
-        && fd.effects.is_empty()
-        && (!facts.builtin_calls.is_empty() || !facts.calls_to.is_empty())
-    {
-        labels.push("trivial-helper");
-    }
-
-    if facts.body_stmt_count == 1
-        && !facts.has_match
-        && !self_call
-        && fd.effects.is_empty()
-        && facts.builtin_calls.is_empty()
-        && facts.calls_to.is_empty()
-        && facts.ctor_constructs.is_empty()
-        && !fd.params.is_empty()
-    {
-        labels.push("pure-expression");
-    }
-
-    if facts.body_stmt_count >= 2
-        && !facts.last_stmt_is_match
-        && fd.effects.is_empty()
-        && (!facts.builtin_calls.is_empty() || !facts.calls_to.is_empty())
-        && !self_call
-    {
-        labels.push("let-pipeline");
-    }
-
-    labels
-}
-
-pub fn type_is_result(t: &Type) -> bool {
-    matches!(t, Type::Named { name, .. } if name == "Result")
-}
-
-pub const ARCHETYPE_ORDER: &[&str] = &[
-    "scc-mutual",
-    "structural-recursion",
-    "match-dispatcher",
-    "pipeline-result",
-    "manual-result-adapter",
-    "renderer-formatter",
-    "match-on-value",
-    "orchestration",
-    "effectful-leaf",
-    "let-pipeline",
-    "constructor-wrapper",
-    "data-as-function",
-    "trivial-helper",
-    "pure-expression",
-];
-
-pub fn primary_label(labels: &[&str]) -> &'static str {
-    for &want in ARCHETYPE_ORDER {
-        if labels.contains(&want) {
-            return want;
-        }
-    }
-    "unclassified"
-}
-
-// ─── Call-graph SCC (multi-node strongly connected components) ───────────────
-
-pub fn compute_sccs(fns: &[&ResolvedFnDef], facts_by_id: &HashMap<FnId, &Facts>) -> HashSet<FnId> {
-    let mut graph: HashMap<FnId, Vec<FnId>> = HashMap::new();
-    let fn_ids: HashSet<FnId> = fns.iter().map(|f| f.fn_id).collect();
-    for fd in fns {
-        if let Some(facts) = facts_by_id.get(&fd.fn_id) {
-            let edges: Vec<FnId> = facts
-                .calls_to
-                .iter()
-                .copied()
-                .filter(|c| fn_ids.contains(c) && *c != fd.fn_id)
-                .collect();
-            graph.insert(fd.fn_id, edges);
-        }
-    }
-    let mut index = 0u32;
-    let mut stack: Vec<FnId> = Vec::new();
-    let mut on_stack: HashSet<FnId> = HashSet::new();
-    let mut indices: HashMap<FnId, u32> = HashMap::new();
-    let mut lowlinks: HashMap<FnId, u32> = HashMap::new();
-    let mut multi_scc: HashSet<FnId> = HashSet::new();
-
-    #[allow(clippy::too_many_arguments)]
-    fn strongconnect(
-        v: FnId,
-        graph: &HashMap<FnId, Vec<FnId>>,
-        index: &mut u32,
-        stack: &mut Vec<FnId>,
-        on_stack: &mut HashSet<FnId>,
-        indices: &mut HashMap<FnId, u32>,
-        lowlinks: &mut HashMap<FnId, u32>,
-        multi_scc: &mut HashSet<FnId>,
-    ) {
-        indices.insert(v, *index);
-        lowlinks.insert(v, *index);
-        *index += 1;
-        stack.push(v);
-        on_stack.insert(v);
-        if let Some(neighbors) = graph.get(&v).cloned() {
-            for w in neighbors {
-                if !indices.contains_key(&w) {
-                    strongconnect(
-                        w, graph, index, stack, on_stack, indices, lowlinks, multi_scc,
-                    );
-                    let lv = *lowlinks.get(&v).unwrap();
-                    let lw = *lowlinks.get(&w).unwrap();
-                    lowlinks.insert(v, lv.min(lw));
-                } else if on_stack.contains(&w) {
-                    let lv = *lowlinks.get(&v).unwrap();
-                    let iw = *indices.get(&w).unwrap();
-                    lowlinks.insert(v, lv.min(iw));
-                }
-            }
-        }
-        if lowlinks.get(&v) == indices.get(&v) {
-            let mut scc: Vec<FnId> = Vec::new();
-            loop {
-                let w = stack.pop().unwrap();
-                on_stack.remove(&w);
-                scc.push(w);
-                if w == v {
-                    break;
-                }
-            }
-            if scc.len() > 1 {
-                multi_scc.extend(scc);
-            }
-        }
-    }
-
-    let nodes: Vec<FnId> = graph.keys().copied().collect();
-    for v in nodes {
-        if !indices.contains_key(&v) {
-            strongconnect(
-                v,
-                &graph,
-                &mut index,
-                &mut stack,
-                &mut on_stack,
-                &mut indices,
-                &mut lowlinks,
-                &mut multi_scc,
-            );
-        }
-    }
-    multi_scc
-}
+pub use crate::analysis::shape::{
+    ARCHETYPE_ORDER, Facts, classify, compute_sccs, extract_facts, primary_label, type_is_result,
+};
 
 // ─── ModuleShape vector + derived Kind ───────────────────────────────────────
